DreamHack/Level5/ChainGallery/requirements.txt
Django==4.2
Pillow==10.0.0  